Changed dorm and flood to pass integer instead of float.

+++ src/com/endicott/edu/simulators/DormManager.java
@@ -87,7 +87,7 @@ public class DormManager {
     }
 
     private void billRunningCostOfDorm(String runId, int hoursAlive, DormitoryModel dorm) {
-        float newCharge = (hoursAlive - dorm.getHourLastUpdated()) * dorm.getMaintenanceCostPerHour();
+        int newCharge = (hoursAlive - dorm.getHourLastUpdated()) * dorm.getMaintenanceCostPerHour();
         Accountant.payBill(runId,"Maintenance of dorm " + dorm.getName() + " $ " + newCharge,(int) (newCharge));
     }
